tools/nitrogen/agents/mac_game_agent/requirements.txt
paho-mqtt
mss
pillow
pyautogui
